Fix mismatch between srand/srand48 and rand/rand48.
Because only the configure macros of spot::srand() had been updated to
the introduction of _config.h, rand() was used even if drand48() was
available, and yet srand48() was being called by spot::srand().  The
consequence is that setting the seed with srand48() had not effect on
the value returned by rand().  Reported by Etienne Renault.

* src/misc/random.cc (drand): Fix configure macros used.